move RPC types to a separate file
