+++ subxt/src/storage/storage_client.rs
@@ -16,7 +16,7 @@ use crate::{
         DecodeWithMetadata,
         Metadata,
     },
-    rpc::{
+    rpc::types::{
         StorageData,
         StorageKey,
     },
