+++ subxt/src/utils/multi_address.rs
@@ -11,8 +11,9 @@ use codec::{
     Encode,
 };
 
-/// A multi-format address wrapper for on-chain accounts. This is a simplified version ob Substrate's
-/// `sp_runtime::MultiAddress`. To obtain more functionality, convert this into that type.
+/// A multi-format address wrapper for on-chain accounts. This is a simplified version of Substrate's
+/// `sp_runtime::MultiAddress`. To obtain more functionality, convert this into that type (this conversion
+/// functionality is provided via `From` impls if the `substrate-compat` feature is enabled).
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Encode, Decode, Debug)]
 pub enum MultiAddress<AccountId, AccountIndex> {
     /// It's an account ID (pubkey).
